Re-implementing USUBSCRIBE command
Updating a subscriber now only involves updating the corresponding row
in the database. Because now we don't remember which deliveries have
occurred in the past we simply don't allow to change any of the
parameters that would affect the list of files associated with a
subscriber (i.e., we don't allow updating the starting date, the
filtering plug-in or the filtering plug-in parameters).

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>