Cache mode set via XML configuration file
The old ways of bringing up a NGAS server in cache mode (via the
ngamsCacheServer/ngamsCacheDaemon scripts for versions < 9.1, using the
-cache command-line option in versions >= 9.1) generate confusion, since
the rest of the cache options are set in the XML configuration file,
alongside the rest of the configuration of an NGAS server. To avoid
future confusions (it has happened in the past) we now moved this option
to the configuration file.

In this new scheme, to start a server in cache mode users need to set
the Caching.Enable flag to 1. If the Caching.Enable flag is set in the
configuration file but the server was *not* started in cache mode, the
server turns the cache mode ON.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>