Adding PID to server's logging format
Now that the server is multiprocess we probably want to be able to
pinpoint to individual processes (and it's anyway always nice to know
the PID).

I'm not so sure whether we will keep this or not; otherwise we can
simply show the PID at the very beginning in a single statement and
don't show it all the time as part of the standard logging.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>