Removed SwVersion handling
Although this had a purpose, it was highly debatable if it did any good
ever. A nicer way to validate a configuration file against a software
version would be to actually validate the XML file against a schema file
shipped with the software, but we're light years away from that.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>