Correctly supporting literal '%' in SQL queries
Our previous implementation didn't take into account that literal '%'
characters in SQL queries do not work in some drivers due to them using
the 'pyformat' or 'format' paramstyles. I haven't double-checked, but we
probably simply didn't have any query in the ngamsDb* modules that used
a literal '%' character, and only now that we added one in the
ngamsDataCheckingThreadTest module the problem became apparent.

This commit fixes the situation and properly replaces the percentage
characters with their escaped version when necessary. We are also now
centralizing the formatting of both the SQL query and the parameters so
we don't need to repeat the same logic in two places.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>