ngamsDbm class cleanup
This includes removing the getRandome method (wasn't being used at all),
removing the unnecesasry pop argument of the getNext method (there is
still one in the "get" method, but that is harder to track down),
removing the usage of TRACE, replacing a Semaphore(1) with a Lock, and
using this lock as a context manager.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>

No need for pop argument

This only made the code more difficult to read (it still is!) while
nobody was using it.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>

dbm cleanup

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>