dumpFileInfoList removed, getFileInfoList improved
The dumpFIL function was being used in two places. In one of them only one
result was needed, while in the other results actually needed to be
stored for later usage (unless we decide to stop supporting the
DbSnapshot feature in NGAS), in which case we still use a bsddb file.

On the other hand, the result of invoking getFIL now is an iterable (a
generator) that retrieves the results from the underlying cursor using
the specified fetch size (and that automatically closes the cursor after
using it). All callers have been updated to the new contract.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>