dumpFileSummary1 removed, getFileSummary1 improved
dumpFS1 dumped the results of getFS1 into a bsddb file, whose name was
then returned to the caller, who in turn needed to open the file to read
the contents, more often than not to write them into a new bsddb file.
This was highly innefficient of course, and therefore a better solution
was needed.

This new implementation fully removes the dumpFileSummary1 method, so
all callers now use the getFileSummar1 method. This method now
internally creates a cursor, and within a closing context manager it
yields the individual results from the cursor. This means that callers
need to simply iterate over the results of this method call without
having to deal with cursors, but getting the benefits of small memory
footprints.

An extra functionality present in dumpFS1 was that it first issued a
COUNT(*) and then a full SELECT using the same criteria, to then compare
the numbers and "verify" that the results were correct. This felt always
like a hack rather than a feature, so I'm simply removing it (if you
can't trust your dbms you better choose a different one!). This means
that now we issue less queries, imposing less load in the database and
speeding up NGAS itself.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>