dumpFileInfo removed, getFIFFI improved
Just like in the previous commit, dumpFI had too much logic in it that
didn't feel correct at all, and also communicated the results back to
the caller by dumping them to a bsddb file on disk. All previous callers
now call getFIFFI (getFileInfoFromFileId), which can either return the
results directly as a sequence, or return a cursor that users can
iterate over.

Users of the dumpFileInfo were the locateArchiveFile (used by RETRIEVE
and other commands), the DISCARD and CLONE commands, and other bits and
pieces, all of which have been adjusted to use getFIFFI.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>