Using cursor2 as context manager
Not only current users of cursor2 have been simplified, but also users
of the old ngamsDbCursor class have been ported to use cursor2 too (and
therefore there is no need to have a use_cursor2 parameter in some of
the methods anymore). Hiding the context management inside the ngamsDb*
classes also allows us to have less indentation on the client code,
which simply needs to loop through the items yielded by the invocation
to the relevant method.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>