DB verify/recover not used anymore
These flags were previosly used to force re-SELECTs, compare the number
of rows returned against a previous COUNT(*), and then issue warnings
and try again until the counts matched. This logic is not present
anymore in the code as we deemed unnecessary to perform these checks in
the first place. Given that, there is also no need for the flags either.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>