Adjusting fabric script to new sample config file
Also, fabric generates a sed command that doesn't work as expected in
MacOS (at least in 10.11, El Capitan) when specifying an empty string as
the backup suffix. The documentation says that this should disable
backup files, but in reality it generates "sed -i -E -e ...", which
creates a backup file with suffice "-E" instead of applying the -E
command line flag (extended regular expressions) to sed, and therefore
failing to find and replace some strings. Our fix is to generate the
backup files, but then remove them.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>