Removing artificial dependency constraint
The constraint imposed over the bsddb3 package was deployment-dependent;
it was there only because in some particular machines we were installing
NGAS on we didn't have the proper combination of the bsddb library and
python packages. I'm removing it now, and users will simply have to be
more careful with what they install (and we can also tweak the fabric
script further, if required).

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>