Safer imports, take 2
The import machinery in python 2.7 puts the modules into the sys.modules
dictionary *before* the modules themselves are fully loaded (which seems
to be to be like this to be able to detect circular dependencies). Thus,
inspecting the sys.modules for the presence of a module must be done
while holding the importing lock, which we are doing now. An alternative
solution would be to not inspect sys.modules and always use importlib's
import_module (or the old __import__ builtin function), but we probably
want to maintain this little optimization, which also enable us to
cleanly offer reloading support.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>