MacOS builds in Travis, take 3
pip commands should be run only once the virtualenv is loaded, which
doesn't happen automatically in MacOS builds. Thus the installation of
the test dependencies happens now in the run_tests.sh script instead of
directly in the .travis.yml file. Likewise, the publication of results
to coveralls also needs the virtualenv loaded, so a new after_success.sh
script has been created to contain the necessary instructions.

In reality we could further separate things, and install all necessary
pip packages in a "before_build.sh" script, only run the installation in
run_build.sh, and only run the tests in run_tests.sh. Still, this
current state of affairs is good enough.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>