MacOS builds in Travis, take 4
The python2 package is already installed in the virtual machines,
alongside with autoconf, automake and libtool. On the other hand "brew
install python2" exits with an error instead of issuing a warning like
it does with the other two packages (probably because there is a newer
version available), so we opt to simply not include any of these
packages as something that needs installing.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>