MacOS builds in Travis, take 6
Setting the hostname via the addons yields a "sudo hostname XYZ", which
seems to work but only so much. In a strange situation I had never
encountered before, python's socket.gethostname() returns different
values (the original hostname, and the new one); the servers spawned by
the test suite get the old name, while *some parts* of the test suite
itself see the new name, and some the old. The difference seems to be
*when* the call is made (import time v/s runtime), but this is only a
very crude conjecture. Anyway, the MacOS hostnames are not that long in
the first place, so there's no need to shorten them artificially.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>