Support for quick server-side failures
When dealing with HTTP communications, the server side can decide to
close the connection quicker than the client expects it to. Before
closing the connection, the server can still write some data back to the
client, probably a small, valid response. Our clients thus should be
aware that this situation can happen, specially when POSTing data. We
thus now recognize when the server is closing the connection while
POSTing, and attempt to read the HTTP response from the server even in
those cases. This allows for quick failures in the server to be seen by
the client quickly as well.

This issue became evident when running the test_ErrHandling_1 test of
the ngamsArchiveCmdTest module in a MacOS environment, but since the
same test never produced the same error in Linux we are now adding a new
test that specifically tests for this scenario.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>