More standard HTTP communications
Some of the unit tests in MacOS environments were failing due to "server
took too long to start" messages, even though the STATUS request was
being seen by the server, and the response sent too. The problem was
that the client was timing out while trying to read the response,
which eventually led me to find that the writing of some responses on
the server side was not correct. In particular, the bytes separating the
HTTP headers from the body were only \n instead of \r\n, and also after
the body was sent an extra \r\n was being written into the response.
This wrong behaviour didn't seem to cause us any harm so far until
testing in MacOS, where it caused intermitent issues. Thus, I've changed
the responses sent by the server to be properly formatted.

Solving that issue exposed a different one in the C client code, where
responses were not being properly parsed, and requests not properly
formatted. Again, this code has been fixed, even though the C client is
usually not used, and reference files have been updated to the correct
expectations.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>