No need to parametrize fixed values
All calls passed on the same values, so we better centralize them
instead of having everyone specify the same. Also, give some love to
that string handling.

This method needs some more attention though, because: a) it should use
the HTTP request's Host header instead of the fqdn of the machine, and
b) because it's probably not necessary at all!

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>