Don't set the Host header
The httplib module will do this for us, and will actually set it to the
correct value (which we were not doing, we were setting it to the
client's hostname).

Also, there's really no need to import the GET and POST constants from a
different module, we can very well do this ourselves.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>