New methods in request handler class
These methods are intended to replace the old and confusing replying
methods of the ngamsServer class. They are easier to identify and to
use, and create errors when users try to send more than one reply back
to the client.

send_file sends a file (from the filesystem) back to the client using
sendfile(2) for better performance, send_data sends arbitrary data of
arbritrary types, which can be read from a bytes string or a read()-able
object. send_status and send_ingest_status are thin wrappers around
send_data, and send an NGAS XML status document. Finally, redirect sends
a redirect reply to the client, and proxy_request acts as a proxy for
the current request, contacting a remote server and returning the
result back to the client.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>