Porting to new send_* methods
The old replying methods of the ngamsServer class (i.e. 'httpReplyGen',
'httpReply', 'httpRedirReply', 'forwardRequest' and 'reply'), which
didn't have meaningful names and were confusing to use, are now gone in
favor of the new methods of the http handler class (i.e., send_file,
send_data, redirect, proxy_request, send_status and send_ingest_status,
plus send_response for body-less responses). This reduced the amount of
code needed in the command modules by quite a lot, making them easier to
maintain in the future and easier to read.

Another simplification made by this commit is that command handling
modules can now simply return a string if they mean to return a success
NGAMS status document back to the client with a particular message. A
few modules are thus now extremely simple, which is again a good thing
to see.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>