Removing unnecessary code
The httpTimeStamp method was only used by the old httpGenReply method of
the ngamsServer. Also, after the porting to the new methods there is no
need to have the ngamsReqProps keep track of the wfile from the request
handler class, or to keep track of whether a reply has been sent or not,
as all writing happens through that class directly, which also keeps
track of its replying state.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>