Reworking NGAS authentication
Authentication in NGAS is supported using HTTP 1.1 Basic authentication.
This new version of the code is more compact, checks for more corner
cases, and uses exceptions to signal unauthenticated or unauthorized
access to resources (and we might actually want to let commands signal
unauthorized access too). Slightly different messages than before are
returned in the HTTP body, which is still an NGAS status document, so
the tests have also been modified, and all reference files usage has
been removed for the sake of simplicity.

In the future if no other clients apart from the old unit tests rely on
the HTTP response body then we can easily change to respond only with
headers using send_response instead of send_status.