Removing all test-related files
The ngamsTest module was a unit test driver, but there is no much
sense on it existing at all as there are much better drivers out there.
In particular we usually use the pytest package to drive the tests.

The ngamsTestPstats module was related to profiling statistics, but
we've never really used it, and looks like the module doesn't implement
much logic anyway. If we want to get some profiling information we can
probably use something more generic though, instead of writing our own
module for that.

Finally, the Makefile.am file was used to compile the C test
application, which has been moved to the ngamsCClient directory,
rendering the Makefile.am unnecessary anymore.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>