Restructuring server startup checks on unit tests
We are still getting some intermittent errors when starting up the
server in MacOS builds. The reasons are not fully clear yet, and part of
that is because we were assuming that all socket.error exceptions were
actually due to connect errors, when in reality they can be thrown due
to something else, like timeouts or more obscure error codes. This
commit re-arranges a bit the startup check sequence to better handle
socket-related errors and properly diagnose them in the future.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>