Proper NGAS server startup check
Since NGAS servers are configured by default to listen only in 127.0.0.1
there is no point on trying to contact them externally to check if they
are up or not. Instead, the easiest way is to reuse what we already have
an invoke 'ngamsDaemon status', which properly handles connection
refusals and a high-level timeout. We don't necessarily want to show an
error and abort the rest of the fabric script if the server is not up
and running, and therefore we simply issue a warning message and
continue.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>