Updating hl.docker_image fabric task
This task had been abandoned for a bit, and it was still producing an
old command line, therefore failing to start containers that ran with
the default command. The command line has been updated to start
correctly.

During this process, several improvements were introduced to the overall
image creation process:

 * The process doesn't fail anymore if the SSH key of the container host
is already present in the local user's known_hosts file. This is more or
less likely to happen, since users can connect to different containers
that at different times use the same IP, and of course we don't want to
fail in those scenarios.

 * The resulting image has been trimmed quite a bit. This is a
combination of *not* installing unnecessary python packages, and of
removing system-level packages that are not necessary at runtime. The
base CentOS 7 image we are using is ~215 [MB], and our resulting image
is ~360 [MB]. The main part of this difference is the virtual
environment with NGAS and all its dependencies at ~140 [MB], of which
numpy and astropy take ~120 [MB].

 * A new option has been introduced to keep the NGAS sources in the
docker image (they are removed by default). This is useful for
development purposes (e.g., when wanting to actually run the tests in an
NGAS container to diagnose problems).

 * Simplified the docker image creation logic. Instead of building a
stage1 image, running a container out of it, and committing the
container into a new image, we now run a container directly from the
base image (centos:centos7), run all commands inside the container, both
directly via docker or indirectly via ssh, and then commit the container
to create the final image. This simplification means that at the end of
the day the whole set of commands run in the container end up being part
of a single filesystem layer, and therefore all the cleanups that we do
at the end also include the cleanup of the initial OpenSSH and sudo
installations, which were previously imprinted on an image filesystem
layer, and therefore increased the size of the final image. Also, with
these changes there is no need to maintain a separate Dockerfile to
build the initial image upon which the target container is started from.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>