Dealing with ECONNRESET errors on connect()
Depending on the OS used at the server side, connect() attempts can come back
with a ECONNRESET error due to the server's connection backlog filling up
quicker than it consumes it, and sending back a RST packet that triggers the
error on the client side (or not sending it, in which case the client will
automatically retry later). Thus, we manually re-try a few times before giving
up hope at the application level.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>