Increase the NGAS request queue backlog
The request queue backlog has a default lenght of 5, which in some highly
concurrent scenarios might not be enough and will create big delays on the
setup of the connection of the concurrent clients. By increasing this size and
matching it to the expect maximum number of simultaneous requests we ensure
that we have a better-sizes backlog queue that fits our needs, and generates
less confusion to our clients.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>