Ensuring cache thread comes up in a timely fashion
A tricky race condition was found during one unit test run, where one of
the ARCHIVE commands finished the bulk of the archiving logic quite
quickly and tried to add the archived file into the cache queue. On the
other hand, the cache control thread took a bit longer to initialize the
objects that implement this queue, leading to the ARCHIVE command
failing due to trying it trying to call "x.y" with x being a None.

With this extra event the threading should tell us when it's safe to
proceed on the main thread, which eventually opens up the door for
external requests to come in and be served.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>