Some datatype adjustments
Now that we are locally testing NGAS against Sybase a few more issues
have been identified regarding the type of the parameters used in some
of the queries (because the driver we use is more pedantic about the
objects being passed around). This commit includes all the adjustments
we found that make Sybase work correctly, but more might come.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>