Support for session-level SQL statements
Whenever a physical connection is established by the connection pool to
the underlying database, one might need to execute one or more SQL
statements to put things in order. For instance, when using the sybpydb
driver (that comes with the Sybase AES installation) there is no option
to connect() to a specific database, and instead users need to issue a
"use XYZ" to switch.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>