Supporting driver-specific named markers
Until now when drivers specify the "named" paramstyle, things have
worked for us if we use ":n1" in the SQL query and "n1" as the key in
the dictionary of values. This scheme however is not mandated by
PEP-249, and drivers can use other marker styles. In particular, when
testing the Sybase driver (to talk to a Sybase ASE server) we found that
the driver needs parameters to be specified like "@n1" in the SQL query,
and also like "@n1" in the dictionary of values.

This driver-dependent knowledge is now part of NGAS, which now uses the
corresponding parameter marker style in the query and in the resulting
data dictionary depending on the driver being used. With time we might
need to add more tweaks for other drivers.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>