Properly calculate this_dir
We always meant this_dir to be an absolute path, so whenever we started
a daemon server (which chdir()'s to /) we would actually get the full
path to the sqlite3 database.

This problems seems to have been hidden by the pytest module loading
mechanism, which caused the correct result to be calculated, but when
using the unittest module things failed miserably.

The fix is obviously easy enough, and now daemon tests are happily
passing when running under unittest.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>