Reusing cursor in transaction object
The previous implementation of the transaction object created a cursor
for each call to execute. This was sub-optimal, since there was no
reason to create the cursor once and use it throughout all the
transaction, which is what we are doing now.

Also, some drivers are unhappy when one calls cursor.execute() with
empty arguments, so we are avoiding these kind of calls now.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>