Create DB objects with smaller pools
Specially during testing there is no need to create pools that require
more than one connection, so let's restrict that. It doesn't mean that
we have been creating more connections than we need so far (if I recall
correctly the steady connections are created on demand), but it will
certainly prevent us from doing it.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>