Slightly improve database connection setup
When loading the server configuration from a database (a very
corner-case scenario, but still supported), we don't actually need to
set the server's database object to the one used to load that
configuration object. In fact, we need a connection pool with a maximum
of one connection to achieve this goal, so we do that instead.

Only after the full configuration has been loaded we proceed to set the
server's database object to the one loaded from the configuration
object, using the configuration's value for the maximum number of
connections in the pool.

This change is simple, but means that we will be able to extend the
PYTHONPATH using the configuration *before* we connect to the database,
which in turn means that we can use database drivers sitting in
external directories, and not necessarily in the NGAS installation
directory.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>