PluginsPath setting becomes more useful
Two main changes happen in this plug-in: on the one hand we now allow a
list of directories to be specified by the PluginsPath setting,
therefore giving more flexibility to users about where external code can
be found. On the other hand, we do this path expansion before loading
the database driver, meaning that users can include the database driver
in the PluginsPath as well, if necessary.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>