Forcefully call _markDeletion
A new environment variable allows users to specify whether they want to
forcefully mark a file for deletion after a successful delivery, but no
delivery count information can be found. In general this shouldn't be
allowed because the file might be needed to perform other deliveries,
but in scenarios where users know there is a single delivery they can
opt to forcefully remove these files.

This is a fix for a problem we found in our MWA on-site deployment,
where files were not being removed. The actual cause of the error (i.e.,
why the file was not found in the fileDeliveryDic structure) is still to
be investigated, but this will allow us to apply a fix without hurting
third parties.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>