Improving sed usage
The sed program in MacOS doesn't recognize the -r flag (which we determined to
be more widely used across Linux distributions), but understands -E, so we
consider that now. It also requires commands given on the command line to be
given with -e, unlike Linux variants, where they are simply understood when
given as a positional argument (but they still understand -e, so it's simply
better to always have -e).

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>