Transitioning to safer imports
This commit modifies all the necessary import statements across our code
base that need to be adjusted in order to be able to import the unit
tests modules using python 3, while retaining successful unit test runs
with python 2. This of course doesn't guarantee that the tests run, or
that *all* our code imports correctly under python 3, but it is a good
start.

Most of the changes consist on correctly importing relative modules.
Python 3 defaults to absolute imports, so trying to import a sibling
module simply with "import sibling_module" fails. On the other hand,
python 2.7 already supports relative imports, so we are using relative
imports now.

Another source of conflict were system modules that were renamed in
python 3. We are using the six.moves mechanism to import those
correctly.

Finally, a few system level modules were already deprecated in python
2.7, and do not exist anymore (e.g., md5 and commands). Already in
python 2.7 there were alternatives for all of them which are still
present in python 3. This commit includes replacing the import with the
newer module, and also the code using those modules.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>