Not using dict.has_key method anymore
The dict.has_key() method has been removed in python 3 in favour of
using the "in" syntax. This is of course also supported in python 2, so
there is no harm on simply replacing all usage of has_key() with the
corresponding "in" or "not in". This also has the nice effect of making
the code easier to read.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>