Proper type checking across python versions
In a few places the code was checking the type of some variables to
distinguish between some basic types (strings, integers, dicts, etc). In
some cases this was being done using type(var) and comparing against
some of the members of the types module, which in others it was being
done using isinstance(var, x), with x also being a member of the types
module. In both cases the problem is the same: the types module doesn't
offer the functionality anymore to distinguish between different base
types. Instead, users should use isinstance(var, x) with x being the
corresponding class (int, str, dict, etc.). In general it is still
correct to use type(var), but this is useful only to distinguish between
more highly different types, like functions, objects and classes for
example.

On top of this, some places were comparing a variable's type against the
basestring class. basestring doesn't exist in python 3 (one simply
checks against str), so we check against six.string_types instead, which
maps to the correct class.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>