Ensuring correct iterables from dictionaries
In python2 some dictionary methods like values() and keys() returned
lists, which is not true anymore in python3. Most of the time this is
OK, since we simply iterate over the results, but some other times we
need the result to be a sequence.

When a sequence is needed from dictionary d, we do "list(d)" to get a
list with the keys of d, and "list(six.itervalues(d))" to get a list
with the values of d. These two changes work in both python 2 and 3, and
are efficient as well in both cases.

Finally, when iterating over the items of a dictionary we do
"d.items()". This works in both python 2 and 3, even though it is
sub-optimal in 2. We however have been doing it like this for a while,
and plan to stop supporting python 2 at some point, so there is no point
in trying to optimize things now.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>