Bytes literals using b'' form
This syntax is accepted by both python 2 and 3, and therefore doesn't
introduce any problems.

In some cases empty literals were being used together with iter() to
detect the end of a data stream in the context of copying data from one
file object and into another. This copying operation is already well
supported in shutils.copyfileobj, so instead of simply fixing the
literal we chose to replace our hand-crafted copying code with the
simpler option of reusing shutils.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>