Opening files in binary mode
Obviously, only when required. We were already doing a pretty good job
at this, but there were a few instances where we were still opening
files without an explicit mode, and therefore python 3 was opening them
in text mode.

In this commit I also took the chance to open files as context managers
when we were not doing so.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>