Produce reliable error messages
When a plug-in is not found, the expected error message contained
whatever came out of str(e), with e being the exception. We are now
taking some control over this situation, properly catching this
particular error and producing our own error message, which will be the
same across python versions.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>