Converting strings to bytes where needed
A few places in the code need bytes instead of strings, so we have to
make sure we perform the appropriate conversion. This is done using
six.b(s), which returns s for python 2, and returns s.encode('latin1')
in python 3. The strings we are converting usually have characters that
live in the ASCII table only, so this strategy is safe enough. If in the
future we need a more flexible way of converting strings to bytes we can
certainly provide one and use it.

Not only we do take care of converting str objects to bytes in python 3
(which is handled using six.b), but we also identify situations where
six.text_type objects (unicode in 2, str in 3) need to be converted to
bytes and react to that.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>