Ensuring lists out of map/filter when required
In python3 the map() and filter() functions don't return a list anymore,
but rather an iterable object. Our code is currently using these results
as lists, so we ensure we always have lists. It might be possible to
change it to use the iterables directly, but that might take some extra
work that lies outside the scope of the work of this branch.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>