Making the crc32 variant work with python3
The crc32 variant (a.k.a. the old ngamsCrc32 checksum plug-in) has
always been unstable. We knew in the past (and there was a comment about
it) that the checksum was going to break with python3, and we
effectively hit this point. This commit makes sure the crc32 variant
works against python 3 as intended.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>