Keeping automatic HTTP body length detection
In python 2.7 the httplib module calculated the length of the body
automatically first based on len(body), if possible, and then assuming
body had a fileno() and could be fstat()-ed to find out its length.
These rules have changed over time, and in python 3.6 they are not there
anymore (a 'Transfer-Encoding: chunked' header is yielded for files, and
len(body) doesn't seem to be used for all objects).

This commit emulates the old behaviour in 3.6+ systems so we can
continue working with the illusion that nothing has changed underneath.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>