Correct header parsing class used
On the one hand we were previously using the Parser class to parse only
MIME headers. This was suboptimal, since python has offered for a while
a HeaderParser class that parses only headers (and thus is simpler and
more efficient). We are using this class now.

On the other hand, python 3 offers two flavours of these classes, one
for parsing string values and other for bytes. Their names are
different, and their relevant method names also differ, so we need to
explicitly use one or the other depending on our python version.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>