Calling self.end_headers explicitly on redirect
In python 3.2+ a call to send_header doesn't necessarily write the
header into the underlying socket, but instead caches them. Due to this,
an explicit call to end_headers must always be done after finishing
writing headers, or after calling send_response without headers.

We were already calling end_headers in most relevant places, except in
the redirect method, which under python 3.6 was failing to send all the
response back to the client.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>