Enforce URLs with a schema
We were previously using the urlopen() method from the urllib module
instead of that from the urllib2 module. The former is quite lenient in
regards to the URLs it accepts, while urllib2's version is more strict
and requires a schema to be always specified.

Python 3 doesn't have the urllib module anymore, so we need to cope with
this situation now. The solution is to simply default to a file://
schema when no schema is specified (which was what urllib's urlopen
defaulted to, too, and what we relied on). In the future we might study
rejecting schema-less URLs.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>