Exercising care with HTTP header handling
The Message class used by the system HTTP-related modules has always
ensured that it allows for case-insensitive header name matching
(e.g., when doing "name in headers" or "del headers[name]"). However,
even though this promise is still true in python 3, the handling of the
actual values of the header names has changed, and now header names
seem to be internally stored in their original form, while in python 2
they were stored in lowercase form.

Our code was implicitly depending in some places in this lowercase-ness
used by the Message class to store header names internally. In some
cases this could be fixed by using the more reliable header name
matching, but in some others (due to the design of our classes) we are
forced to perform multiple-case lookups. Hopefully we can change this in
the future without much effort.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>