Correcting comparissons
This was a big bug that was not apparent until trying the code with
python 3. It turned out that we were comparing, in some places, a string
value against a float, which in python 2 silently performs some kind of
type casting, while in python 3 it throws a TypeError. In another place
we were also comparing a None against a string, which again works in
python 2 but doesn't in python 3. These comparisons still seem to work
in our favour (we haven't seen any problems in our production
environments, and our unit tests always succeed), but we should be
really comparing apples to apples here.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>