Improving ngamsServer command line handling
Previously ngamsServer used a home-brewed command line parser. In the
era of the argparse module this is not necessary anymore, and meant that
we needed to maintain yet more code unnecessarily. Since argparse
already gives us nice help and error messages there is also no need to
maintain a separate text file with the usage help for the ngamsServer
command.

One feature that the old hand-written parser supported though was
handling command line argument names case-insensitvely. This is not
standard behavior in normal applications though, so even though we
maintain this feature, we are explicitly deprecating it and warning
users that they should move away from case-insensitive command line
argument names. We define lower-case only argument names in the argparse
parse, and internally we have moved all our invocations to use these
lower-case argument names.

Also significantly, argument parsing previously occurred *within* the
ngamsServer class, while now it happens outside. During this process the
constructor and the init() method of the ngamsServer class changed,
which broke some of its subclasses used by the unit tests. I thus took
the opportunity to clean them up and make them as light as possible. In
the future I plan by adding a -serverclass command-line option that
accepts a classname to use as the server class. This will allow to put
most of these classes into a single module instead of having a single
module for each, which seems otherwise unnecessary.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>